changed Amount filed type of History Entry to double

+++ src/Lykke.Service.OperationsHistory.Core/Entities/HistoryLogEntryEntity.cs
@@ -14,13 +14,13 @@ namespace Lykke.Service.OperationsHistory.Core.Entities
             return Guid.NewGuid().ToString();
         }
         public DateTime DateTime { get; set; }
-        public decimal Amount { get; set; }
+        public double Amount { get; set; }
         public string Currency { get; set; }
         public string OpType { get; set; }
         public string ClientId { get; set; }
         public string CustomData { get; set; }
 
-        public static HistoryLogEntryEntity Create(DateTime dateTime ,decimal amount, string currency, string clientId, string customData, string opType)
+        public static HistoryLogEntryEntity Create(DateTime dateTime, double amount, string currency, string clientId, string customData, string opType)
         {
             return new HistoryLogEntryEntity
             {
